Testnet 0

+++ sss.asado-ledger/src/main/scala/sss/asado/balanceledger/BalanceLedger.scala
@@ -3,8 +3,10 @@ package sss.asado.balanceledger
 import java.util
 
 import sss.ancillary.Logging
+import sss.asado.AsadoEvent
 import sss.asado.account.NodeIdentity
-import sss.asado.block.BlockId
+import sss.asado.chains.Chains.GlobalChainIdMask
+import sss.asado.common.block.BlockId
 import sss.asado.contract.LedgerContext._
 import sss.asado.contract.{CoinbaseDecumbrance, CoinbaseValidator, LedgerContext, SinglePrivateKey}
 import sss.asado.identityledger.IdentityService
@@ -12,9 +14,11 @@ import sss.asado.ledger._
 import sss.asado.util.ByteArrayEncodedStrOps._
 import sss.db.Db
 
+import scala.util.Try
+
 object BalanceLedger {
-  def apply(cbe : CoinbaseValidator,identityService: IdentityService)(implicit db:Db) : BalanceLedger =
-    new BalanceLedger(new UTXODBStorage(), cbe, identityService)
+  def apply(cbe : CoinbaseValidator,identityService: IdentityService)(implicit db:Db, chainIdMask: GlobalChainIdMask) : BalanceLedger =
+    new BalanceLedger(new UTXODBStorage(chainIdMask), cbe, identityService)
 }
 
 trait BalanceLedgerQuery {
@@ -33,6 +37,7 @@ class BalanceLedger(storage: UTXODBStorage,
 
   override def entry(inIndex: TxIndex): Option[TxOutput] = storage.get(inIndex)
 
+
   /**
     * Used for debug only.
     *
@@ -44,7 +49,35 @@ class BalanceLedger(storage: UTXODBStorage,
 
   def keys: Seq[TxIndex] = storage.keys
 
-  def apply(stx: SignedTxEntry, blockHeight: Long) {
+  private def validateCoinbase(blockHeight: Long,
+                               sigs: Seq[Seq[Array[Byte]]],
+                               tx: Tx,
+                               inputs: Seq[TxInput]): Int = {
+
+    require(inputs.size == 1, "Coinbase can only have 1 input")
+
+    val in = inputs.head
+    in.txIndex match {
+      case TxIndex(coinbaseTxId, 0) if CoinbaseTxId sameElements coinbaseTxId =>
+
+        log.debug(s"Coinbase validation for height $blockHeight")
+
+        coinbaseValidator.validate(blockHeight, sigs, tx)
+        // No need to delete from storage because it's not in storage.
+          if (blockHeight % 100 == 0) {
+          log.info(s"Balance ledger balance is ${balance} at height $blockHeight, " +
+          s"adding ${in.amount} via coinbase")
+        }
+        in.amount
+
+      case TxIndex(_, _) =>
+        throw new IllegalArgumentException(s"Only one coinbase tx allowed ${in.txIndex}.")
+    }
+  }
+
+  private[balanceledger] def apply(stx: SignedTxEntry, blockId:BlockId): Unit = {
+
+    val blockHeight = blockId.blockHeight
 
     val tx = stx.txEntryBytes.toTx
     storage.inTransaction {
@@ -55,32 +88,33 @@ class BalanceLedger(storage: UTXODBStorage,
       require(stx.signatures.length == ins.length, "Every input *must* have a specified param sequence, even if it's empty.")
       var totalIn = 0
 
-      ins.indices foreach { i =>
-        val in = ins(i)
-        entry(in.txIndex) match {
-          case Some(txOut) =>
-            require(txOut.amount == in.amount, s"In amount ${in.amount}) must equal out amount ${txOut.amount} ")
-            totalIn += in.amount
-            val context = LedgerContext(contextTemplate + (blockHeightKey  -> blockHeight))
-            require(txOut.encumbrance.decumber(txId +: stx.signatures(i), context, in.sig), "Failed to decumber!")
-            storage.delete(in.txIndex)
+      if (blockId.txIndex == 1) {
+        //this is coinbase
+        totalIn += validateCoinbase(blockHeight, stx.signatures, tx, ins)
 
-          case None => in.txIndex match {
-            case TxIndex(coinbaseTxId, 0) if CoinbaseTxId sameElements coinbaseTxId =>
+      } else {
+        ins.indices foreach { i =>
+          val in = ins(i)
+          entry(in.txIndex) match {
+            case Some(txOut) =>
+
+              require(txOut.amount == in.amount,
+                s"In amount ${in.amount}) must equal out amount ${txOut.amount} "
+              )
+
               totalIn += in.amount
-              log.info(s"Coinbase validation for height $blockHeight")
-              coinbaseValidator.validate(blockHeight, stx.signatures, tx)
-              // No need to delete from storage because it's not in storage.
-              if(blockHeight % 100 == 0) {
-                log.info(s"Balance ledger balance is ${balance} at height $blockHeight, adding ${in.amount} via coinbase")
-              }
+              val context = LedgerContext(contextTemplate + (blockHeightKey -> blockHeight))
 
-            case TxIndex(coinbaseTxId, _) if CoinbaseTxId sameElements coinbaseTxId =>
-              throw new IllegalArgumentException(s"Only one coinbase tx allowed ${in.txIndex}.")
-            case _ => throw new IllegalArgumentException(s"${in.txIndex} does not exist.")
+              require(txOut.encumbrance.decumber(txId +: stx.signatures(i), context, in.sig),
+                "Failed to decumber!"
+              )
+              storage.delete(in.txIndex)
+
+            case None => throw new IllegalArgumentException(s"${in.txIndex} does not exist.")
           }
         }
       }
+
 
       var totalOut = 0
       outs.foldLeft(0) { (acc, out) =>
@@ -91,30 +125,41 @@ class BalanceLedger(storage: UTXODBStorage,
       }
 
       require(totalOut == totalIn, s"Total out (${totalOut}) *must* equal total in (${totalIn})")
-
     }
   }
 
   override def balance: Int = storage.entries.foldLeft(0)((acc, e) => acc + e.amount )
 
-  override def apply(ledgerItem: LedgerItem, blockHeight: Long): Unit = {
+  override def apply(ledgerItem: LedgerItem, blockId: BlockId): LedgerResult = Try {
     val stx = ledgerItem.txEntryBytes.toSignedTxEntry
-    require(util.Arrays.equals(stx.txId, ledgerItem.txId), s"The transmitted txId (${ledgerItem.txId.toBase64Str}) does not match the generated TxId (${stx.txId.toBase64Str})")
-    apply(stx, blockHeight)
+    require(util.Arrays.equals(stx.txId, ledgerItem.txId),
+      s"The transmitted txId (${ledgerItem.txId.toBase64Str}) " +
+        s"does not match the generated TxId (${stx.txId.toBase64Str})"
+    )
+
+    apply(stx, blockId)
+
+    Seq.empty
   }
 
-  override def coinbase(nodeIdentity: NodeIdentity, blockId: BlockId, ledgerId: Byte): Option[LedgerItem] = {
+  override def coinbase(nodeIdentity: NodeIdentity, forBlockHeight: Long, ledgerId: Byte): Option[LedgerItem] = {
 
     if(coinbaseValidator.rewardPerBlockAmount > 0) {
       val coinbaseAmount = coinbaseValidator.rewardPerBlockAmount
       val minNumBlocksWait = coinbaseValidator.numBlocksInTheFuture
-      val ins: Seq[TxInput] = Seq(TxInput(TxIndex(CoinbaseTxId, 0), coinbaseAmount, CoinbaseDecumbrance(blockId.blockHeight)))
+      val ins: Seq[TxInput] = Seq(TxInput(TxIndex(CoinbaseTxId, 0),
+        coinbaseAmount,
+        CoinbaseDecumbrance(forBlockHeight))
+      )
 
-      val outs = Seq(TxOutput(coinbaseAmount, SinglePrivateKey(nodeIdentity.publicKey, blockId.blockHeight + minNumBlocksWait)))
+      val outs = Seq(TxOutput(coinbaseAmount,
+        SinglePrivateKey(nodeIdentity.publicKey, forBlockHeight + minNumBlocksWait))
+      )
       val cb = StandardTx(ins, outs)
       val txSignature = nodeIdentity.sign(cb.txId)
-      val stx = SignedTxEntry( cb.toBytes, Seq(Seq(txSignature)))
+      val stx = SignedTxEntry(cb.toBytes, Seq(Seq(txSignature)))
       Some(LedgerItem(ledgerId, stx.txId, stx.toBytes))
+
     } else None
   }
 }
